requirements.txt
paddlepaddle-gpu
paddleocr>=2.0.1
opencv-python
pillow